Fix all the compile error related to new parser, typeck fails still
