+++ stuff/parse_ui/misc/maths.cm
@@ -1,14 +1,12 @@
-int add(int x, int y) { return x + y; }
+fn add(int x, int y) -> int { return x + y; }
 
-void main() {
-    int x, y, z, a, b;
-
-    z = 4 >> 2 & 0 << 3;
-    x = 4 + 2 & 0 + 3;
+fn main() {
+    let z = 4 >> 2 & 0 << 3;
+    let x = 4 + 2 & 0 + 3;
     z += 1;
-    y = z | (1 & ~5);
+    // let y = z | (1 & ~5);
     z = y + add(1, 1);
-    a = 5;
+    let a = 5;
     a += 5;
-    b = a + 1;
+    let b = a + 1;
 }
